Refactor: Move file utility functions to lib/file-utils
